feat(git api deserialization fix & parallel link scrapping & tests): github api rework event deser ; add parallel link parsing to scrapper ; add tests for new logic of description formatting and prallel parsing

+++ scrapper/src/main/java/backend/academy/linktracker/scrapper/property/ScrapperGlobalProperties.java
@@ -30,8 +30,9 @@ public class ScrapperGlobalProperties {
 
     int parallelScrappersAmount = 5;
 
-    long maxDescriptionLength = 200;
+    int maxDescriptionLength = 200;
 
     @NotNull
-    Instant defaultLinkCreationDate = Instant.EPOCH; // Not min, but epoch cause cant format to date string in request it
+    Instant defaultLinkCreationDate =
+            Instant.EPOCH; // Not min, but epoch cause cant format to date string in request it
 }
